рефрактор
