+++ src/view/event.js
@@ -1,5 +1,5 @@
 import {
-  changeRirstLetteToUpperCase,
+  changeFirstLetteToUpperCase,
   formatToShortDay,
   formatToDateTime,
   formatToFullDateTime,
@@ -41,7 +41,7 @@ const createEventTemplate = (event = {}) => {
         <img class="event__type-icon" width="42" height="42" src="img/icons/${type}.png" alt="Event type icon">
       </div>
       <h3 class="event__title">
-        ${changeRirstLetteToUpperCase(type)}
+        ${changeFirstLetteToUpperCase(type)}
         ${destination.name}
       </h3>
       <div class="event__schedule">
